% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ches1217}
\alias{ches1217}
\title{ches1217 dataset}
\format{A graph (iGraph object or list) with this configuration:
\describe{
\item{Vertices}{Career program. It includes 10 oder features used in the orginal paper.}
\item{Edges}{Binary link between programs.}
}}
\usage{
ches1217
}
\description{
Graph of Chilean Higher Education Space between 2012-2017.
}
\examples{
data(ches1217)
summary(ches1217)
}
\references{
Candia, C., Encarnação, S., & Pinheiro, F. L. (2019). The higher education space: Connecting degree programs from individuals’ choices. EPJ Data Science, 8(1), 1–17. https://doi.org/10.1140/epjds/s13688-019-0218-4
}
\keyword{dataset}
