% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_classification}
\alias{get_classification}
\title{Get Data of students clasified}
\usage{
get_classification(n = 0)
}
\description{
Load data of students. Filter only enrolled and situation(P), then classify among dropped and not dropeed.
}
\examples{

}
