% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Get Data dir}
\usage{
get_data_dir()
}
\value{
Return a string with the path to a found data directory.
}
\description{
Retrieves the location of data, searching several possible directories.
}
