% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data_postulaciones}
\alias{get_data_postulaciones}
\title{Get Data of  Postulaciones}
\usage{
get_data_postulaciones()
}
\value{
a dataframe with postulations.
}
\description{
Imports data of type .Rda prevously created and saved in local pc. Also import demre.csv by using function get_demre.
}
\examples{

}
