% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_demre}
\alias{get_demre}
\title{Get demre}
\usage{
get_demre()
}
\description{
Load data from demre.csv. It is a file previously curated manually. It includes BigGroup field and Stem Field.
}
\examples{

}
