% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_stem}
\alias{get_stem}
\title{Get Stem}
\usage{
get_stem(demre)
}
\arguments{
\item{demre}{a dataframe previously got with function get_demre()}
}
\value{
Information of Stem degree
}
\description{
Using demre data, previously computed, indicates Stem programs.
}
\examples{

}
