% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plot_hes}
\alias{plot_hes}
\title{Plot Network}
\usage{
plot_hes(g)
}
\arguments{
\item{g}{A graph object (igraph graph).}
}
\description{
Plots a network of kind Higher Educational space.
}
\examples{
plot_hes(g=ches0611)
}
