% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_alumno_coe}
\alias{get_alumno_coe}
\title{Get Alumno coe}
\usage{
get_alumno_coe(g = ches1217, classification, data_postulaciones)
}
\description{
Alumnos merged with postulaciones merged with coherence.
}
\examples{

}
