% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data_postulaciones}
\alias{get_data_postulaciones}
\title{Get Data of  Postulaciones}
\usage{
get_data_postulaciones()
}
\description{
Imports data of type .Rda prevously created and saved in local pc. Also import demre.csv
}
\examples{
<<<<<<< HEAD

=======
postulaciones <- get_data_postulaciones()
>>>>>>> 2ba6505fcde673ae3749badc403009fe888de9e3
}
